Bug-000 Remove repeated mouseclick events
As I plan to explain in the Pull Request: I experienced a runtime bug
where the *same mouseclick was processed several times in a row.*
Based on my knowledge writing games in SDL2 in the past, I wrapped the
PollEvent call inside a while-loop in order to resolve it -- and sure
enough, it worked! On my machine (a Linux machine using GCC 10), each
mouseclick was registered "once and only once," no matter how long the
mouse button was "held down."

But, this update does come with a disclaimer. I have modified this
code to use my local libraries, which are SDL2 libraries rather than
SDL3. Thus, if this change does not work (or does not noticeably
change any of the runtime logs) on *your* machine, please do not merge
it!! It could be an outdated approach.

Either way, I am obliged to say, thank you so much for the
opportunity. It was great to "scratch the itch" and try running
another SDL game after so many years away from it.